✨feat: Task 생성
https://github.com/CarSpotterHQ/server/issues/16

+++ src/main/java/com/carspotter/CarSpotter/repository/MajorCategoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MajorCategoryRepository extends JpaRepository<MajorCategory, Integer> {
+public interface MajorCategoryRepository extends JpaRepository<MajorCategory, Long> {
     List<MajorCategory> findByNameContaining(String name);
 }
